navbar headaches
