+++ package.json
@@ -11,7 +11,8 @@
     "ethers": "^5.4.4",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
-    "react-scripts": "3.2.0"
+    "react-scripts": "5.0.0",
+    "react-bootstrap": "^2.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
