+++ src/App.css
@@ -28,14 +28,17 @@
 .waveButton {
   cursor: pointer;
   margin-top: 16px;
-  border: 1px solid rgb(23, 58, 94);
-  border-radius: 10px;
+  float: right;
+  margin-left: 200px;
+  border: 1px solid rgb(112, 216, 112);
+  border-radius: 20px;
   color: rgb(255, 255, 255);
-  background: rgb(0, 30, 60);
+  background: rgb(112, 216, 112);
   width: 200px;
-  padding: 16px;
-  margin: 20px auto;
+  padding: 6px;
+  margin: 2px auto;
   text-transform: uppercase;
+  font: 2em;
 }
 
 input {
@@ -65,6 +68,7 @@ input {
 .twitter:hover {
   border-bottom: 1px dashed rgb(135, 126, 220);
 }
+
 
 .message {
   color: rgb(255, 255, 255);
