+++ src/App.js
@@ -2,126 +2,60 @@ import React, { useEffect, useState } from "react";
 import { ethers } from "ethers";
 import './App.css';
 import abi from "./utils/abi.json"
-import axios from "axios";
-
+import Header from "./components/header.js";
 
 
 const App = () => {
+
   const [currentAccount, setCurrentAccount] = useState("");
   const [items, setItems] = useState([]);
-
 
   /*
    * All state property to store all waves
    */
 
-  /*
-   * Create a method that gets all waves from your contract
-   */
-
-  const getData = async () => {
-    if (currentAccount) {
-     try {
-      const api_endpoint = `https://api.covalenthq.com/v1/1/address/${currentAccount}/transactions_v2/?quote-currency=USD&format=JSON&block-signed-at-asc=false&no-logs=false&page-size=10000&key=ckey_8f68249e3fce44109f7386dabc5:`
-      console.log(api_endpoint)
-      const response = await axios.get(api_endpoint);
-      console.log(response.data);
-      setItems(response.data.data.items)
-         } catch (err) {
-      // Handle Error Here
-      console.error(err);
-     }} else {
-       console.error("NOT Connected");
-     }
-}
-
-  
-  
   const checkIfWalletIsConnected = async () => {
-    try {
+      try {
       const { ethereum } = window;
 
       if (!ethereum) {
-        console.log("Make sure you have metamask!");
-        return;
+          console.log("Make sure you have metamask!");
+          return;
       } else {
-        console.log("We have the ethereum object", ethereum);
+          console.log("We have the ethereum object", ethereum);
       }
 
       const accounts = await ethereum.request({ method: 'eth_accounts' });
 
       if (accounts.length !== 0) {
-        const account = accounts[0];
-        console.log("Found an authorized account:", account);
-        setCurrentAccount(account);
+          const account = accounts[0];
+          console.log("Found an authorized account:", account);
+          setCurrentAccount(account);
       } else {
-        console.log("No authorized account found")
+          console.log("No authorized account found")
       }
-    } catch (error) {
+      } catch (error) {
       console.log(error);
-    }
-  }
-
-  /**
-  * Implement your connectWallet method here
-  */
-  const connectWallet = async () => {
-    try {
-      const { ethereum } = window;
-
-      if (!ethereum) {
-        alert("Get MetaMask!");
-        return;
       }
-
-      const accounts = await ethereum.request({ method: "eth_requestAccounts" });
-
-      console.log("Connected", accounts[0]);
-      setCurrentAccount(accounts[0]); 
-    } catch (error) {
-      console.log(error)
     }
-  }
-
-  
 
   useEffect(() => {
     checkIfWalletIsConnected();
   }, [])
   
   return (
-    <div className="mainContainer">
-      <div className="dataContainer">
-        <div className="header">
-        👋 Hey there!
-        </div>
+    <div>
+      <Header items={items} currentAccount={currentAccount}
+        setCurrentAccount={setCurrentAccount} setItems={setItems}
+      />
+      <div className="mainContainer">
 
-        <div className="bio">
-          Check out your CO2 emissions in Web3!
-        </div>
-        <button className="waveButton" onClick={getData}>
-          Get Climate FootPrint
-        </button>
+        <div className="dataContainer">
 
-
-        
-        {/*
-        * If there is no currentAccount render this button
-        */}
-        {!currentAccount && (
-          <button className="waveButton" onClick={connectWallet}>
-            Connect Wallet
-          </button>
-        )}
-
-        {currentAccount && (
-          
           <div className="bio">
-          Transactions: {items.length}
+            Check out your CO2 emissions in Web3!
           </div>
-        )}
-
-        
+        </div>
       </div>
     </div>
   );
